Make 3dchess executable compile on MSVC with /WX
- Tons of double->float errors resolved
- Replaced float and double versions of OpenGL functions with
  integer version where only integer arguments were given
- Realized that static_cast<float>'ing the world is annoying and
  that (float) is much shorter and not very dangerous in this
  project.
- Set _CRT_SECURE_NO_WARNINGS and _SCL_SECURE_NO_WARNINGS to get
  rid of errors not in our code or not fixable in a portable way.

+++ src/gui/AnimationHelper.cpp
@@ -39,7 +39,7 @@ float AnimationHelper::ease(FunctionType type, const float lowerBound, const flo
 
 	switch (type) {
 		case EASE_OUTSINE:
-			m_completeness = sinf(elapsed_time / static_cast<float>(m_duration)* (M_PI / 2));
+			m_completeness = sinf(elapsed_time / static_cast<float>(m_duration) * (static_cast<float>(M_PI) / 2.f));
 			break;
 		case EASE_LINEAR:
 		default:
